updated jQuery, Bootstrap; general syntax updates
Updated jQuery to v1.11.3, Bootstrap to v3.3.4. interface.php has been
updated to accommodate the changes in Bootstrap on the home page.
Further work is necessary to update the remaining portions of the file.
Readme has been updated with a demo site. Remaining files are edited as
a result of a minor reformatting in Phpstorm.