Enhanced mail function.
A number of new options are now configurable from config.php. It is now
possible to define the senders address and display name from config.php.
Mail format has been changed from plaintext to html format.
In addition, it is now possible to use the PHPmailer() library if
installed.

The new config options are :

- $fromEmail : Will be used as default sender address if not overriden by
  authenticated user.

- $fromName : Will be used as the display name of fromEmail

- $signature :  A text string to be added as signature to the email. This
  Should be written as HTML.

- $PHPMailer  : Enables or disables the use of PHPMailer

- $PHPMailerPath : Path to PHPMailers src folder

- $PHPMailerSmtp : Use SMTP rather than standard mail() with PHPMailer

- $PHPMailerHost : SMTP host to use if SMTP is enabled

- $PHPMailerPort : Port for SMTP host

- $PHPMailerSecure : Set to SSL/TLS/StartTLS or blank

- $PHPMailerUser : Username to authenticate against SMTP server

- $PHPMailerPassword : Password for given SMTP user